fim do modulo

+++ poo/Interfaces/ICalculadora.cs
@@ -10,6 +10,6 @@ namespace poo.Interfaces
         int Somar(int num1, int num2);
         int Subtrair(int num1, int num2);
         int Multiplicar(int num1, int num2);
-        double Dividir(int num1, int num2);
+        double Dividir(double num1, double num2);
     }
 }
